[returns] Allow a matrix for 'weights'
'weights' may now also be a matrix with as many rows
as 'rebalance.when'. Also, rebalancing at the first
timestamp is no longer implied, i.e. a function
invocation such as

    returns(x, weights = <...>,
            rebalance.when = 2)

will only rebalance at the second timestamp, and the
first return will be zero. To get back the old
behaviour, say

    returns(x, weights = <...>,
            rebalance.when = c(1, 2))